Don't clobber msginfo register after invocation
The messageinfo register is set to 0 after processing a kernel object
invocation if the invocation set the thread to restart.

This is a problem if the kernel object invocation had also set the
message info register.

The existing convention in this case is for the kernel invocation to set
the thread state to running right before it returns to avoid the
problematic code path.

However, there are some invocations that do not follow this convention
and their message info register gets clobbered. The affected kernel
object invocations are:

  - ARMPageGetAddress
  - RISCVPageGetAddress
  - X86PageGetAddress
  - X86VCPUReadVMCS
  - SchedContextConsumed
  - SchedContextYield

This commit fixes these.

Signed-off-by: Jimmy Brush <code@jimmah.com>